Add files via upload
Habib's talks eventually showed me the importance of e-portfolios and resumes. To hire the top individuals, companies will look at portfolios and resumes for proof of their skills.
While listening to PPG's talk, I realized that if a data engineering student wants to be a data engineer, they must strengthen their basic coding and programming skills, such as MySQL and Java, because they will be working with big data later on.