Add files via upload
Design Thinking Project taught me some new skills which is how to communicate and deal with people. Besides that, I also have a better understanding of databases because my topic is database systems. I also learned data administrators handle data and monitor it, some of the software that they use is Oracle 19c and MySQL.