Add files via upload
The Credence talk on Webex states that on July 6, 2022, Telekom Malaysia Berhad (TM) introduced Credence. Credence focuses on cloud and digital services to help the public sector and businesses with their digital transformation efforts, as well as its relationships with VMware, AWS, and Huawei. I learned the importance of technical expertise, problem-solving skills, lifelong learning, and good communication in order to become a future system developer.