Add files via upload
This slide includes our prototype which is PowerBM. Our prototype is about the software that has a built-in AI, NISA AI that can help the user sort their file or data according to what the user wants. Our software also has many features which is data visualization and normalisation (beta version).  Our software can generate a chart or visualize the data that they receive through the NISA AI. Users also can request from the built-in AI anything as long as the request is under the control of the AI.