Add files via upload
This video is about the UTM Digital visit. UTM Digital is the Department of Digital Services for UTM. In this UTM Digital is formed with 6 other components which is DigitalCare, MyUTM, UTMSmart, Services, Data Centre, and Awareness Program. During this visit, I've go 3 places in UTM Digital which are Open Office, UTM Digital Care, and Data Centre, but the most thing that I learned was from the data center. UTM's Data Center services more than 350 physical servers and accommodates up to 2520U size of physical server which is 840U of size in use. I also learned how they operate the Data and how frequence they manage the data.